requirements.txt
gradio
flask
gunicorn
requests
pillow
pytube